requirements.txt
click
google-cloud-monitoring
rich
pync

google-api-python-client

pyyaml